Adding move operations to mutable erasure.
The immutable erasure leverages std::shared_ptr semantics to be copyable and movable. The mutable erasure cannot trivially adopt either move or copy semantics: mutable copying implies deep copying to create unique instances; moving requires handling non-movable types indirectly. The mutable erasure implements a configurable small object optimization to store objects internally. However, that is not viable for non-movable objects without rendering the erasure itself immovable.

The movability issue is resolved by only allowing movable types to be used with the small object storage space. Immovable types are stored via std::make_unique, using the same facilities for large objects. Having branched model type basedsize/movability, movability is implemented via a virtual function that uses placement new to invoke the model's move constructor in external storage, passing the current model as the argument.

Two suboptimal points: 1) the model looses sense of whether it is empty, since there is no generic interface to interrogate wheather it has been moved from. While more performant to not require state, it also makes it more difficult to interrogate the erasure meaningfully--the model type remains the same before/after move. We could modify the erasure's move implementation to construct an empty model after move, but that introduces virtual calls into the path that add no meaning aside from the uncommon case of interrogtation. Likewise model state would bloat storage. In the future it could be made into a configuration point, but for now it's left in a performant/permissive externally opaque state.

2) It's non-trivial to implement moving from one size of erasure to another. Principally, the internal/external storage model decision becomes dynamic. It's still pretty straight forward to implement, but requires implementing a mixed size move/copy path.

It would also be nice to use type-safe methods for move/copy (discard void* for std::aligned_storage), however, currently getting different alignment deductions depending on where in the virtual hierarchy I apply alignof. Probably something that should be probed and corrected.

Lastly, a copy operator should be provided on a best-effort basis. This is a bit more complex in that copyability cannot be imbued into an object--we can make an object movable by adding indirection, but not copy it. This implies more of a throw-if-not-copyable approach, regardless of storage location. Ideally, it would be easier to specify copy/move as noexcept, and integrate into that exception pathway. However, since we're not in a noexcept-by-default langauage, that's probably trying to be too generous.

Perhaps it could be a configuration point?